스케쥴러 저장 기능 추가

+++ payout-service/src/main/java/com/mchung/payoutservice/entity/DailyRevenue.java
@@ -28,7 +28,7 @@ public class DailyRevenue {
     private Long advertisementId;
 
     @Column(name="totalRevenue")
-    private Long totalRevenue;
+    private Double totalRevenue;
 
     @Column(name="effectiveDate")
     private Date effectiveDate;
